moved generateMarkdown.js to a new subfolder
